+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/Utility/HDUtils.cs
@@ -460,7 +460,15 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                     SystemInfo.graphicsDeviceType == GraphicsDeviceType.Vulkan ||
                     SystemInfo.graphicsDeviceType == (GraphicsDeviceType)22 /*GraphicsDeviceType.Switch*/);
         }
-        
+
+        public static void CheckRTCreated(RenderTexture rt)
+        {
+            // In some cases when loading a project for the first time in the editor, the internal resource is destroyed.
+            // When used as render target, the C++ code will re-create the resource automatically. Since here it's used directly as an UAV, we need to check manually
+            if (!rt.IsCreated())
+                rt.Create();
+        }
+
 #if UNITY_EDITOR
         // This function can't be in HDEditorUtils because we need it in HDRenderPipeline.cs (and HDEditorUtils is in an editor asmdef)
         public static bool IsSupportedBuildTarget(UnityEditor.BuildTarget buildTarget)